update parcel build pipeline
