+++ src/layout/components/index.ts
@@ -1,5 +1,5 @@
 import { packEnclose, packSiblings } from 'd3-hierarchy'
-import { Node, Edge, connectedComponents } from '../../'
+import { Node, Edge, connectedComponents } from '../../trellis'
 
 export type Options = Partial<{
   padding: number
