+++ src/layout/hierarchy/index.ts
@@ -1,5 +1,5 @@
 import { hierarchy, HierarchyPointNode, tree } from 'd3-hierarchy'
-import { Node, Edge } from '../../'
+import { Node, Edge } from '../../trellis'
 
 export type Options = Partial<{
   x: number
