+++ src/renderers/webgl/interaction/drag.ts
@@ -1,6 +1,6 @@
 import * as PIXI from 'pixi.js-legacy'
 import { InternalRenderer } from '..'
-import { Node, Edge } from '../../..'
+import { Node, Edge } from '../../../trellis'
 import { clientPositionFromEvent } from '../utils'
 
 /**
