+++ src/renderers/webgl/sprites/arrowSprite.ts
@@ -1,6 +1,6 @@
 import * as PIXI from 'pixi.js-legacy'
 import { InternalRenderer } from '..'
-import { Node, Edge } from '../../..'
+import { Node, Edge } from '../../../trellis'
 
 export class ArrowSprite<N extends Node, E extends Edge> {
   static ARROW_HEIGHT = 12
